+++ src/com/company/cardGame/blackJack/Actor.java
@@ -1,15 +1,15 @@
 package com.company.cardGame.blackJack;
 public interface Actor {
-    byte HIT = 1;
-    byte STAND = 2;
-    byte DOUBLE = 3;
-    byte SPLIT = 4;
-    byte QUIT = 0;
+    int HIT = 1;
+    int STAND = 2;
+    int DOUBLE = 3;
+    int SPLIT = 4;
+    int QUIT = 0;
 
     String getName();
     int getBalance();
     int placeBet();
-    byte getAction(Hand hand, Hand dealer);
+    int getAction(Hand hand, Hand dealer);
     void addBalance(double amt);
     //byte getAcion(Hand hand, List<Cards> cards) -> card counter version
 }
